feat: override bloc add handler (#3)
Adds an override to the add handler for StateMachine.

Since the handlers for these events happen in the periphery of Bloc -
and aren't handled inside of `on` event handlers, they're little more
than external functions adding events to the bloc.

If these were in `on` handlers, the bloc would know about the ongoing
event handlers (or be able to await them) and allow them to complete
before fully closing the Bloc.

This could probably be removed if you moved StateMachine to instead
handle events via an `on` handler after they are defined. But that's for
another day.

This change simply ignores events added after `close` is called, but
before any ongoing state emits are pending. This prevents a `StateError`
from bubbling up from async side effects that are still processing if
the bloc is fully closed.